requirements.txt
crewai
crewai-tools
python-dotenv
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google_auth_oauthlib
pydub
markdown
streamlit 
fpdf 
python-docx 
PyPDF2
reportlab
markdown2
weasyprint
html2docx